replace rewiring tasks with doLast() call to fix Gradle 9 compatibility
